+++ vsproject/vsproject/utilities.h
@@ -9,9 +9,9 @@
 int load_image(
     char* name
 ,   std::vector<float> &img
-,   unsigned * width
-,   unsigned * height
-,   unsigned * chnls
+,   size_t * width
+,   size_t * height
+,   size_t * chnls
 );
 
 //! Write image
